productsmanagement/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
python-decouple==3.8
py-eureka-client==0.13.2